+++ src/Components/Chat/ChatLayout.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react';
 import io from 'socket.io-client';
-import { USER_CONNECTED, LOGOUT } from '../Events';
+import { USER_CONNECTED, LOGOUT } from '../../Events';
 import LoginForm from './LoginForm';
 
 const socketUrl = "http://localhost:3231";
